requirements.txt
requests==2.31.0
pytz==2023.3
pyodbc==5.1.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
pytest==8.4.2